test-amr-monitoring/requirements.txt
paho-mqtt
requests
keyboard